Squashed commit of the following:
commit ad2d4ccb2329c6ee0c262d96676ab8926c073748
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Mon Jan 31 23:48:32 2022 -0600

    made posting automatic

commit fc6aa2b27742a96812cdf0021a83da50e3771ab3
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Mon Jan 31 22:55:57 2022 -0600

    punched up service layer tests for comments and likes

commit 59e3bd6299c006b90bf6e35c0d010fa7f3f2f647
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Mon Jan 31 21:41:32 2022 -0600

    redundant test

commit 7bce40d57743e726fd65b78c6b7eaec575eeae5c
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Mon Jan 31 21:31:41 2022 -0600

    updated the page to include the new search bar, added logout

commit fde916dc410a7fe0de34616688ff12b0912b578c
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Mon Jan 31 19:28:49 2022 -0600

    added the old css to the css file

commit c7e6ec99437e01177dcdb9ce7d0018fc4341396d
Author: ericthered1138 <ericjennings955@gmail.com>
Date:   Mon Jan 31 18:56:37 2022 -0600

    recentered the page, there was one </div> out of place

+++ PythonAPI/service_layer/implementation_classes/postfeed_service.py
@@ -1,3 +1,5 @@
+from custom_exceptions.post_id_must_be_an_integer import PostIdMustBeAnInteger
+from custom_exceptions.user_id_must_be_an_integer import UserIdMustBeAnInteger
 from data_access_layer.implementation_classes.postfeed_dao import PostFeedDaoImp
 from service_layer.abstract_classes.postfeed_service_abs import PostfeedService
 
@@ -11,7 +13,15 @@ class PostFeedServiceImp(PostfeedService):
         return self.post_feed_dao.get_all_posts()
 
     def delete_a_post_service(self, post_id: int):
+        # Check to make sure the post_id is an integer
+        if not str(post_id).isnumeric():
+            raise PostIdMustBeAnInteger('The post id must be an integer.')
+
         return self.post_feed_dao.delete_a_post(post_id)
 
     def get_all_posts_by_user_id_service(self, user_id: int):
+        # Check to make sure the user_id is an integer
+        if not str(user_id).isnumeric():
+            raise UserIdMustBeAnInteger('The user id must be an integer.')
+
         return self.post_feed_dao.get_all_posts_with_user_id(user_id)
